SuperC: functionAnalysis: optimizations, static func check
If two function definition/call share the same presence condition, functionAnalysis
can be optimized by processing the presence condition for only one and using
the processed for the other directly. Do this optimization to greatly improve
the performance for functionAnalysis.

Do distinguish static function definitions from global ones, and have an
option to ignore definitions for and calls to static functions to improve
performance in exchange with not analysing the static functions.